Moving methods to a new resolver for unauthorized user access
